fix a lot
