+++ apps/open-swe/src/graphs/manager/nodes/index.ts
@@ -1,4 +1,4 @@
 export * from "./initialize-github-issue.js";
-export * from "./classify-message.js";
+export * from "./classify-message/index.js";
 export * from "./start-planner.js";
 export * from "./create-new-session.js";
